SW-16169 - Add missing order detail attributes upon adding order position in backend
When adding a new order position via the shopware backend, there are currently no attributes written to the database for that new position. This might lead to problems, if a plugin relies on order detail attributes (table `s_order_details_attributes`) being present. This commit adds the missing creation of a `Shopware\Models\Attribute\OrderDetail` instance, when adding a new order position.

closes #655