SW-16344 - Set order number in order details upon converting a cancelled order
When converting a cancelled order to a *normal* order, the newly determined order number is currently not set in the order details. This behaviour diverges from the order creation during the checkout process. This commit fixes this behaviour by setting the new order number in all order details too.

This commit does not introduce any breaking changes.