Fix PHP 7 fatal error in mPDF SVG rendering
When rendering HTML to a PDF using mPDF, it is generally possible that the HTML contains SVG tags. However, when running PHP 7, a fatal error can be raised while rendering the SVG: `Fatal error: 'break' not in the 'loop' or 'switch' context`. As a result, it is impossible to use mPDF for rendering any HTML containing SVG in PHP 7. This commit fixes that fatal error by replaying the wrong `break` with `return`. The same fix was done in the main mPDF repository too (https://github.com/mpdf/mpdf/commit/002bb8ae97f81299ac79e472a86c051993e43444).