SW-19160 - Use ‘ocramius/proxy-manager’ in Enlight_Hook_ProxyFactory fixes https://github.com/shopware/shopware/pull/1174
When `Enlight_Hook_ProxyFactory` generates a proxy class, all hooked methods of the original class are implemented, defining all parameters defined in the original method signatures. However, only the parameter and optionally a default value are currently read from the original signature and added to the proxy method signature. That is, any parameter type hints are not copied over. This works for most cases, but results in a fatal error when hooking a class that implements an interface, which defines a method with type-hinted parameters. The PHP docs for interfaces state this behaviour in a note: 'The class implementing the interface must use the exact same method signatures as are defined in the interface. Not doing so will result in a fatal error.'
This commit improves `Enlight_Hook_ProxyFactory` by using the `ocramius/proxy-manager` library to generate the proxy classes, which takes care of everything from resolving namespaces to generating exact copies of method signatures.
Please note that we use version `^1.0` of `ocramius/proxy-manager`, because `v2` is not compatible with PHP 5 anymore.

Related to shopware/shopware#1065 and shopware/shopware#1159