SW-20968 - Fix Uncaught TypeError when passing null to addFacet() method
Since Shopware allows to open the details of an article that is mapped to inactive categories only an "Fatal error: Uncaught TypeError: Argument 1 passed to Shopware\Bundle\SearchBundle\ProductNumberSearchResult::addFacet() must implement interface Shopware\Bundle\SearchBundle\FacetResultInterface, null given, called in /var/www/shop/htdocs/engine/Shopware/Bundle/SearchBundleES/FacetHandler/CategoryFacetHandler.php on line 145 and defined in /var/www/shop/htdocs/engine/Shopware/Bundle/SearchBundle/ProductNumberSearchResult.php:87" occurs when trying to search for the article.

Because the buildFacetResult() method of the CategoryTreeFacetResultBuilder returns TreeFacetResult or null it's possible that a $facet of type NULL is going to be injected to the ProductNumberSearchResult's addFacet() method.

To avoid that error it's necessary to wrap the addFacet() call inside a instanceof FacetResultInterface condition.
fixes #1473