SW-21445 - fix dot in parameter key for parameters added by QueryBuilder->addFilter on cygwin php7
The more_entropy parameter of uniqid defaults to true in php7 compiled on cygwin / from the cygwin package repository. Therefor $parameterKey contains a dot in this case which leads to an invalid query on evaluation.
This just includes the return value of uniqid in the str_replace which is already done to remove dots from the expression key.
fixes #1576